2011-05-13 17:32 UTC+0200 Viktor Szakats (harbour.01 syenar.hu)   * contrib/hbwin/hbwin.ch   * contrib/hbwin/win_svc.c     + added 4th param to WIN_SERVICEINSTALL() to spec start type     + added related constants
  * contrib/hbwin/wapi_misc.c
    ! cast for win64

+++ harbour/contrib/hbwin/wapi_misc.c
@@ -113,7 +113,7 @@ static TCHAR * hbwapi_FileNameAtSystemDir( const TCHAR * pFileName )
       LPTSTR buffer;
 
       if( pFileName )
-         nLen += hbwapi_tstrlen( pFileName ) + 1;
+         nLen += ( UINT ) hbwapi_tstrlen( pFileName ) + 1;
 
       buffer = ( LPTSTR ) hb_xgrab( nLen * sizeof( TCHAR ) );
